:sparkles: Allow finishing the run if the population collapses
This is pretty much impossible given how this app works right now, but it
might be fun to let a user type in their own starting string, in which case
it could happen right away.

Also, I'm thinking there might be an interesting mode where we model a
simple form of abiogenesis, where we keep throwing random strings at the
problem until one kicks off an explosion of duplication and mutation.